backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# CORS
fastapi-cors==0.0.6

# Data Processing & ML
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2

# Explainable AI
shap==0.46.0

# Database
psycopg2-binary==2.9.10
python-dotenv==1.0.1

# API Client
pydantic==2.9.2
pydantic-settings==2.6.1

# Text Processing
unicodedata2==15.1.0

# Validation
email-validator==2.2.0
